refactor: split index file
