+++ src/shared/generate.ts
@@ -1,5 +1,3 @@
-import type { WireguardResponse, XrayResponse } from "./types/conn";
-
 export function generatePassword(length = 12): string {
   const chars =
     "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()_+{}:\"<>?|[];',./`~";
@@ -9,12 +7,6 @@ export function generatePassword(length = 12): string {
     password += chars[idx];
   }
   return password;
-}
-
-export function isXrayResponse(
-  proto: WireguardResponse | XrayResponse,
-): proto is XrayResponse {
-  return (proto as XrayResponse).Xray !== undefined;
 }
 
 export function generateUsername(totalLength = 25): string {
