+++ src/types/env.d.ts
@@ -1,0 +1,8 @@
+declare namespace NodeJS {
+  interface ProcessEnv {
+    BOT_TOKEN: string;
+    API_BASE_URL: string;
+    API_AUTH_TOKEN: string;
+    DOMAIN: string;
+  }
+}
